refactor(tests, bitvector): add static casts

+++ src/bitvector.cpp
@@ -15,7 +15,7 @@ uint8_t countOneBits(size_t n) {
 Bitvector::Bitvector(std::string bits)
 : bitvector(bits.size() / 64 + (bits.size() % 64 == 0 ? 0 :  1)),
   size(bits.size()),
-  rankBlockSize(std::max(floor(log2(bits.size())/2), 1.0)), //< type warnings not really important here. Expected to be below 64
+  rankBlockSize(static_cast<size_t>(std::max(floor(log2(static_cast<double>(bits.size()))/2), 1.0))), //< type warnings not really important here. Expected to be below 64
   rankSuperblockSize(rankBlockSize * rankBlockSize),
   rankSuperblocks(bits.size() / rankSuperblockSize + (bits.size() % rankSuperblockSize == 0 ? 0 : 1)),
   rankBlocks(bits.size() / rankBlockSize + (bits.size() % rankBlockSize == 0 ? 0 : 1)),
